Make FileTree a struct wrapping the actual tree.
This makes the names a bit less hokey (`treeify()` -> `FileTree::new()`)
and the path -> metadata function a nice method.
It seems a little silly to have a wrapper struct around the actual tree,
but we'll make the root directory `pub` for easy access.

The Rayon example and a few other docs are TODO,
but otherwise things are starting to look a lot more usable.