Make FileTree a trait to extend functionality to any directory
For any directory in a file tree, we'd like to be able to lookup a child
file (or directory) by path, or iterate through all child
files/directories/both.

Putting the methods in a struct that wraps the top level (like we were)
keeps us from performing these actions on child directories, so that's
a no go.

We can't add them as methods to `struct Directory`, since we don't
wrap the top-level entries in some phony root directory
(what would its `Metadata` be?). So that's a no go.

Since our file tree is a nested set of BTreeMaps,
our only real option is to add a trait and then impl it for the maps.

Since the methods will extend BTreeMap, let's use `lookup()` instead of
`get()`, and `traverse()` instead of `iter()`.